Finished the CRUD template for categories Added alerts - Sweetalert

+++ pages/index/categories.vue
@@ -6,7 +6,10 @@
         <div class="row">
           <div class="col-lg-12">
             <div class="block margin-bottom-sm">
-              <div class="title"><strong>Food Categories</strong></div>
+              <div class="title">
+                <strong>Food Categories</strong>
+                <button class="transparent-button material-icons float-right" @click="editModal(false)" data-toggle="modal" data-target="#categoryModal">add</button>
+              </div>
               <div class="table-responsive">
                 <table class="table">
                   <thead>
@@ -22,6 +25,7 @@
                       <td>{{ category.name }}</td>
                       <td>
                         <button type="button" @click="editModal(true); selectCategory(category.id)" data-toggle="modal" data-target="#categoryModal" class="btn btn-primary">Edit</button>
+                        &nbsp;
                         <button type="button" @click="deleteCategory(category.id)" class="btn btn-danger">Delete</button>
                       </td>
                     </tr>
@@ -30,49 +34,39 @@
               </div>
             </div>
           </div>
-
-          <!-- Modal Form -->
-          <div class="col-lg-4">
-            <div class="block">
-              <div class="title"><strong>Create a new category</strong></div>
-              <div class="block-body text-center">
-                <button type="button" @click="editModal(false)" data-toggle="modal" data-target="#categoryModal" class="btn btn-primary">New Category</button>
-                <!-- Modal-->
-                <div id="categoryModal" tabindex="-1" role="dialog" aria-hidden="true" class="modal fade text-left">
-                  <div role="document" class="modal-dialog">
-                    <div class="modal-content">
-                      <div class="modal-header">
-                        <strong class="modal-title" v-if="!isEditModal">New Category</strong>
-                        <strong class="modal-title" v-else>Update Category</strong>
-                        <button type="button" @click="closeModal()" aria-label="Close" class="close"><span aria-hidden="true">×</span></button>
-                      </div>
-                      <div class="modal-body">
-                        <p v-if="!isEditModal">Add a new food category.</p>
-                        <p v-else>Update the category.</p>
-                        <form v-if="!isEditModal" @submit.prevent="createCategory()">
-                          <div class="form-group">
-                            <label>Name</label>
-                            <input v-model="category.name" type="text" placeholder="Name" class="form-control" required>
-                          </div>
-                          <div class="form-group">
-                            <input type="submit" value="Create" class="btn btn-primary">
-                          </div>
-                        </form>
-                        <form v-else @submit.prevent="updateCategory()">
-                          <div class="form-group">
-                            <label>Name</label>
-                            <input v-model="category.name" type="text" placeholder="Name" class="form-control" required>
-                          </div>
-                          <div class="form-group">
-                            <input type="submit" value="Update" class="btn btn-primary">
-                          </div>
-                        </form>
-                      </div>
-                      <div class="modal-footer">
-                        <button type="button" @click="closeModal()" class="btn btn-secondary">Close</button>
-                      </div>
+          <!-- Modal-->
+          <div id="categoryModal" tabindex="-1" role="dialog" aria-hidden="true" class="modal fade text-left">
+            <div role="document" class="modal-dialog">
+              <div class="modal-content">
+                <div class="modal-header">
+                  <strong class="modal-title" v-if="!isEditModal">New Category</strong>
+                  <strong class="modal-title" v-else>Update Category</strong>
+                  <button type="button" @click="closeModal()" class="transparent-button material-icons">close</button>
+                </div>
+                <div class="modal-body">
+                  <p v-if="!isEditModal">Add a new food category.</p>
+                  <p v-else>Update the category.</p>
+                  <form v-if="!isEditModal" @submit.prevent="createCategory()">
+                    <div class="form-group">
+                      <label>Name</label>
+                      <input v-model="category.name" type="text" placeholder="Name" class="form-control" required>
                     </div>
-                  </div>
+                    <div class="form-group">
+                      <input type="submit" value="Create" class="btn btn-primary">
+                    </div>
+                  </form>
+                  <form v-else @submit.prevent="updateCategory()">
+                    <div class="form-group">
+                      <label>Name</label>
+                      <input v-model="category.name" type="text" placeholder="Name" class="form-control" required>
+                    </div>
+                    <div class="form-group">
+                      <input type="submit" value="Update" class="btn btn-primary">
+                    </div>
+                  </form>
+                </div>
+                <div class="modal-footer">
+                  <button type="button" @click="closeModal()" class="btn btn-secondary">Close</button>
                 </div>
               </div>
             </div>
@@ -87,6 +81,7 @@
 <script>
   import axios from 'axios'
   import ContentHeader from '~/components/ContentHeader.vue'
+  const alertHandling = require('~/static/js/alert-handling.js')
 
   export default {
     components: {
@@ -110,6 +105,7 @@
             this.categories = categories.data
           })
           .catch(err => {
+            alertHandling.error(err.response.data.message)
             this.errors.push(err)
           })
       },
@@ -119,9 +115,10 @@
           .then(category => {
             this.categories.push(category.data)
             this.closeModal()
+            alertHandling.success("Your new category was created")
           })
           .catch(err => {
-            this.closeModal()
+            alertHandling.error(err.response.data.message)
             this.errors.push(err)
           })
       },
@@ -133,6 +130,7 @@
             this.category = category.data
           })
           .catch(err => {
+            alertHandling.error(err.response.data.message)
             this.errors.push(err)
           })
       },
@@ -142,29 +140,36 @@
           .then(category => {
             for (let i = 0; i < this.categories.length; i++) {
               if (this.categories[i].id === category.data.id) {
-                this.categories.splice(i, 1)
+                this.categories[i] = category.data
               }
             }
-            this.categories.push(category.data)
             this.closeModal()
+            alertHandling.success("Your category was updated")
           })
           .catch(err => {
+            alertHandling.error(err.response.data.message)
             this.errors.push(err)
           })
       },
 
       deleteCategory(categoryId) {
-        axios.delete(process.env.apiUrl + "/categories/" + categoryId)
-          .then(response => {
-            for (let i = 0; i < this.categories.length; i++) {
-              if (this.categories[i].id === categoryId) {
-                this.categories.splice(i, 1)
+        let that = this
+        alertHandling.confirm("Are you sure you want to delete this category?", function() {
+          axios.delete(process.env.apiUrl + "/categories/" + categoryId)
+            .then(response => {
+              for (let i = 0; i < that.categories.length; i++) {
+                if (that.categories[i].id === categoryId) {
+                  that.categories.splice(i, 1)
+                  alertHandling.success("Your category was removed")
+                  break;
+                }
               }
-            }
-          })
-          .catch(err => {
-            this.errors.push(err)
-          })
+            })
+            .catch(err => {
+              alertHandling.error(err.response.data.message)
+              that.errors.push(err)
+            })
+        })
       },
 
       editModal(isEditModal) {
